requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-telegram-bot==20.7
asyncio==3.4.3
Werkzeug==2.3.7 